gen_check/requirements.txt
asgiref==3.8.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.3
djangorestframework==3.15.1
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
pdfkit==1.0.0
pillow==10.2.0
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
